Move db-related types into the specific api filtering module
 * move JoinedTables and Action out of ApiOperations
 * separate Operations Filtering execution into composable steps as
   for other similar implementors
 * remove unused classes and imports
